pyproject.toml
[project]
name = "remark-drop"
version = "0.1.0"
description = "Push Twitter/X articles to your Kindle or reMarkable"
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "readability-lxml>=0.8.1",
    "python-dotenv>=1.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "html2text>=2024.2.26",
    "requests>=2.31.0",
    "weasyprint>=61.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[tool.setuptools]
py-modules = ["main", "extractor", "emailer", "dropbox_uploader"]
